Add login support
Login is managed via oauth. Only if
1. The user successfully logs in
2. Is part of the whitelist

Are they allowed into the site.
The whitelist is hardcoded in the source but allows wildcards, eg:
`*@t4gc.com`.

Also adds a LoginView component and the related background image.
Login user info is stored in browser local storage and all urls are
guarded with the login check.