Fix login refactoring to hide file upload
Previously because of a v-if mess up the file upload would persist into
the dashboard view blocking the search bar. This commit un-reverts the
login changes, and adds a v-if-else to avoid this issue.

This commit is built atop a "Revert "Revert "Add login support"""
commit, so the only change is the `v-if-else` added to the login page.