Center on points not drawn polygon
Previously we were using the use drawn polygon to center the map.
Since there might not be a user drawn polygon in cases without drawing
(i.e. pre-loaded data) we center based on the points themselves now.

This still has the desired effect of filtering the datapoints because
the WriterMap still passes the savedGeoJson layers to the Dataviewer,
which removes the unnecessary points before going into the dashboard panel.

For backward compatibility we retain the savedGeoJson behavior.